BLIP/requirements.txt
albumentations==1.3.0
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
contourpy==1.0.7
cycler==0.11.0
fairscale==0.4.4
filelock==3.11.0
fonttools==4.39.3
ftfy==6.1.1
huggingface-hub==0.13.4
idna==3.4
imageio==2.28.0
importlib-resources==5.12.0
joblib==1.2.0
kiwisolver==1.4.4
lazy_loader==0.2
matplotlib==3.7.1
mkl-fft==1.3.1
mkl-random==1.2.2
mkl-service==2.4.0
networkx==3.1
numpy==1.23.5
olefile==0.46
opencv-python==4.7.0.72
opencv-python-headless==4.7.0.72
packaging==23.0
Pillow==9.5.0
pip==23.0.1
pycocoevalcap==1.2
pycocotools==2.0.6
pyparsing==3.0.9
python-dateutil==2.8.2
PyWavelets==1.4.1
PyYAML==6.0
qudida==0.0.4
regex==2023.3.23
requests==2.28.2
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
sacremoses==0.0.53
safetensors==0.3.1
scikit-image==0.20.0
scikit-learn==1.2.2
scipy==1.9.1
setuptools==65.6.3
six==1.16.0
some-package==0.1
threadpoolctl==3.1.0
tifffile==2023.4.12
timm==0.4.12
tokenizers==0.10.3
torch==1.10.0
torchaudio==0.10.0
torchvision==0.11.0
tqdm==4.65.0
transformers==4.15.0
typing_extensions==4.5.0
urllib3==1.26.15
wcwidth==0.2.6
wheel==0.38.4
zipp==3.15.0
